Breakout SymbolDrawable
